Update travis deployment key
Updating the key is error-prone and requires the secret key. It
also requires the travis command line utility. Using that and
the key I was finally able to update .travis.yml using

travis encrypt -x -a deploy.key -r leamas/bsb4_pi  <my key>

-x: Overwrite existing key
-r: Name of repo

Note that the repo is part of the key so after forking a new,
encrypted key must be generated.